Refactor code through changing bidirectional association to unidirectional association
